Users and Admin Added
Not only have the views, viewmodels and controllers been added, but the admin role has been created, and I have created an account that fills the role of admin.
Next: I need to add authorization/authentication annotiations for controller methods that need them, and I need to relate data (such as user and comments) to the top-level domain models.