Ready For Migration
Removed legacy embedded views and controller. Updated the scaffolded controller 'embeds' to include user. Re-organized embed>details to be more aesthetically pleasing.

I need to get the list of comments incorporated into each tool, but a comment will need a common identifier to be associated with any given tool. As it stands, each class that inherits from tool has it's own primary key, but tool itself does not have a primary key. I'm going to try and migrate a primary key into tool, and use that. Better design would have obviated the other primary keys, but here we are. Who knows, it might even come in handy to have a secondary or composite key.