PK-Refactor Pre-Migration
I need to have only one PK so that I can use comments for each tool more effectively. I'm going to use tool to do it, as it's the root entity.

I've deleted most of the other scaffolds, and done some reworking on the models.

I'm ready to migrate then update.