PK-Refactor Pre-DB-Update
Lots of stuff went wrong. I caught and corrected several errors that stemmed from the refactor, to the point that the project could be built. It is ready to go.

The DB looks janky, again, but this should allow me to access the user of a comment on any tool.