PK-Refactor Post-DB-Update
Wow! It works, and it works well!
I was wondering how it was going to handle the different types of tools, and it decided to opt in to just having one table called 'Tools' with a 'Discriminator' column that sorts out what type of tool it is. The discriminator column is followed by nullable columns containing the fields specific to the models associated with the discriminated tool.
Neat.