Tool is now associated with user
The embedsController has been altered so that the user signed in creating a tool  has their username assigned to and associated with the tool they're adding to the site. This is different from the tool's creator, which is another name for author.

Next task: Get comment box to display correctly.