Gave up on comments, starting on remaining domain controllers.
I believe the correct way out of my delima of associating comments with a particular guide is view components, but I don't have time to learn what they are and how to create them.
Instead, I'm cutting the comment feature, and moving forward with scaffolding the rest of the models to finish up this project.