Getting Started
I'm going to be adding unit tests, but the best documentation to do so for this platform exists within Murach's MVC textbook.
The first roadblock I've run into is that Murach uses repositories in the constructors of his controllers, whereas mine are set up with the context.
I'll need to do some refactoring so that the constructors use the IUnitOfWork class rather than the context. Once I have that, I can model my unit tests after those found in the text.